refactor: extract shared AWS type validators into carina-aws-types crate
Create a new carina-aws-types crate containing ~25 type validators that were
duplicated between carina-provider-aws and carina-provider-awscc. Both providers
now re-export from the shared crate via `pub use carina_aws_types::*`.

Shared validators include: resource ID types (vpc_id, subnet_id, etc.), ARN
validators (with service-specific validation), enum helpers, IPAM pool ID,
AWS account ID, KMS key ID, availability zone validation, tags_type, and
VALID_REGIONS constant.

Provider-specific types remain in their respective crates: region validators
(different namespaces), schema config structs, and AWSCC's IAM policy document
validation.

Closes #553
