+++ Cargo.toml
@@ -1,5 +1,6 @@
 [workspace]
 members = [
+    "carina-aws-types",
     "carina-cli",
     "carina-codegen-aws",
     "carina-core",
