+++ carina-aws-types/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "carina-aws-types"
+version = "0.1.0"
+edition = "2024"
+license = "MIT"
+
+[lib]
+doctest = false
+
+[dependencies]
+carina-core = { path = "../carina-core" }
