+++ carina-aws-types/src/lib.rs
@@ -1,0 +1,995 @@
+//! Shared AWS type definitions and validators
+//!
+//! This module contains type validators shared between `carina-provider-aws`
+//! and `carina-provider-awscc`. Provider-specific types (region with namespace,
+//! IAM policy document, schema config structs) remain in their respective crates.
+
+use carina_core::resource::Value;
+use carina_core::schema::AttributeType;
+
+// ========== Enum helpers ==========
+
+/// Check if `input` matches any of `valid_values` using enum matching rules:
+/// exact match, case-insensitive, or underscore-to-hyphen (case-insensitive).
+/// Returns the matched valid value if found.
+pub fn find_matching_enum_value<'a>(input: &str, valid_values: &[&'a str]) -> Option<&'a str> {
+    // Exact match
+    if let Some(&v) = valid_values.iter().find(|&&v| v == input) {
+        return Some(v);
+    }
+    // Case-insensitive match
+    if let Some(&v) = valid_values
+        .iter()
+        .find(|&&v| v.eq_ignore_ascii_case(input))
+    {
+        return Some(v);
+    }
+    // Underscore-to-hyphen match (case-insensitive)
+    let hyphenated = input.replace('_', "-");
+    if let Some(&v) = valid_values
+        .iter()
+        .find(|&&v| v.eq_ignore_ascii_case(&hyphenated))
+    {
+        return Some(v);
+    }
+    None
+}
+
+/// Canonicalize an enum value by matching against valid values.
+/// Handles exact match, case-insensitive match, and underscore-to-hyphen conversion.
+pub fn canonicalize_enum_value(raw: &str, valid_values: &[&str]) -> String {
+    find_matching_enum_value(raw, valid_values)
+        .unwrap_or(raw)
+        .to_string()
+}
+
+// ========== Region constants ==========
+
+/// Valid AWS regions (in AWS format with hyphens)
+pub const VALID_REGIONS: &[&str] = &[
+    "ap-northeast-1",
+    "ap-northeast-2",
+    "ap-northeast-3",
+    "ap-southeast-1",
+    "ap-southeast-2",
+    "ap-south-1",
+    "us-east-1",
+    "us-east-2",
+    "us-west-1",
+    "us-west-2",
+    "eu-west-1",
+    "eu-west-2",
+    "eu-west-3",
+    "eu-central-1",
+    "eu-north-1",
+    "ca-central-1",
+    "sa-east-1",
+];
+
+// ========== Tags ==========
+
+/// Tags type for AWS resources (map of string values)
+pub fn tags_type() -> AttributeType {
+    AttributeType::Map(Box::new(AttributeType::String))
+}
+
+// ========== Resource ID validators ==========
+
+/// Validate a generic AWS resource ID format: `{prefix}-{hex}` where hex is 8+ hex digits.
+pub fn validate_aws_resource_id(id: &str) -> Result<(), String> {
+    let Some(dash_pos) = id.find('-') else {
+        return Err("expected format 'prefix-hexdigits'".to_string());
+    };
+
+    let prefix = &id[..dash_pos];
+    let hex_part = &id[dash_pos + 1..];
+
+    if prefix.is_empty()
+        || !prefix
+            .chars()
+            .all(|c| c.is_ascii_lowercase() || c.is_ascii_digit())
+    {
+        return Err("prefix must be lowercase alphanumeric".to_string());
+    }
+
+    if hex_part.len() < 8 {
+        return Err("ID part must be at least 8 characters after prefix".to_string());
+    }
+
+    if !hex_part.chars().all(|c| c.is_ascii_hexdigit()) {
+        return Err("ID part must contain only hex digits".to_string());
+    }
+
+    Ok(())
+}
+
+/// Validate a resource ID with a specific prefix (e.g., "vpc", "subnet").
+pub fn validate_prefixed_resource_id(id: &str, expected_prefix: &str) -> Result<(), String> {
+    let expected_format = format!("{}-xxxxxxxx", expected_prefix);
+    if !id.starts_with(&format!("{}-", expected_prefix)) {
+        return Err(format!("expected format '{}'", expected_format));
+    }
+    validate_aws_resource_id(id)
+}
+
+/// AWS resource ID type (e.g., "vpc-1a2b3c4d", "subnet-0123456789abcdef0")
+#[allow(dead_code)]
+pub fn aws_resource_id() -> AttributeType {
+    AttributeType::Custom {
+        name: "AwsResourceId".to_string(),
+        base: Box::new(AttributeType::String),
+        validate: |value| {
+            if let Value::String(s) = value {
+                validate_aws_resource_id(s)
+                    .map_err(|reason| format!("Invalid resource ID '{}': {}", s, reason))
+            } else {
+                Err("Expected string".to_string())
+            }
+        },
+        namespace: None,
+        to_dsl: None,
+    }
+}
+
+/// VPC ID type (e.g., "vpc-1a2b3c4d")
+pub fn vpc_id() -> AttributeType {
+    AttributeType::Custom {
+        name: "VpcId".to_string(),
+        base: Box::new(AttributeType::String),
+        validate: |value| {
+            if let Value::String(s) = value {
+                validate_prefixed_resource_id(s, "vpc")
+                    .map_err(|reason| format!("Invalid VPC ID '{}': {}", s, reason))
+            } else {
+                Err("Expected string".to_string())
+            }
+        },
+        namespace: None,
+        to_dsl: None,
+    }
+}
+
+/// Subnet ID type (e.g., "subnet-0123456789abcdef0")
+pub fn subnet_id() -> AttributeType {
+    AttributeType::Custom {
+        name: "SubnetId".to_string(),
+        base: Box::new(AttributeType::String),
+        validate: |value| {
+            if let Value::String(s) = value {
+                validate_prefixed_resource_id(s, "subnet")
+                    .map_err(|reason| format!("Invalid Subnet ID '{}': {}", s, reason))
+            } else {
+                Err("Expected string".to_string())
+            }
+        },
+        namespace: None,
+        to_dsl: None,
+    }
+}
+
+/// Security Group ID type (e.g., "sg-12345678")
+pub fn security_group_id() -> AttributeType {
+    AttributeType::Custom {
+        name: "SecurityGroupId".to_string(),
+        base: Box::new(AttributeType::String),
+        validate: |value| {
+            if let Value::String(s) = value {
+                validate_prefixed_resource_id(s, "sg")
+                    .map_err(|reason| format!("Invalid Security Group ID '{}': {}", s, reason))
+            } else {
+                Err("Expected string".to_string())
+            }
+        },
+        namespace: None,
+        to_dsl: None,
+    }
+}
+
+/// Internet Gateway ID type (e.g., "igw-12345678")
+pub fn internet_gateway_id() -> AttributeType {
+    AttributeType::Custom {
+        name: "InternetGatewayId".to_string(),
+        base: Box::new(AttributeType::String),
+        validate: |value| {
+            if let Value::String(s) = value {
+                validate_prefixed_resource_id(s, "igw")
+                    .map_err(|reason| format!("Invalid Internet Gateway ID '{}': {}", s, reason))
+            } else {
+                Err("Expected string".to_string())
+            }
+        },
+        namespace: None,
+        to_dsl: None,
+    }
+}
+
+/// Route Table ID type (e.g., "rtb-abcdef12")
+pub fn route_table_id() -> AttributeType {
+    AttributeType::Custom {
+        name: "RouteTableId".to_string(),
+        base: Box::new(AttributeType::String),
+        validate: |value| {
+            if let Value::String(s) = value {
+                validate_prefixed_resource_id(s, "rtb")
+                    .map_err(|reason| format!("Invalid Route Table ID '{}': {}", s, reason))
+            } else {
+                Err("Expected string".to_string())
+            }
+        },
+        namespace: None,
+        to_dsl: None,
+    }
+}
+
+/// NAT Gateway ID type (e.g., "nat-12345678")
+pub fn nat_gateway_id() -> AttributeType {
+    AttributeType::Custom {
+        name: "NatGatewayId".to_string(),
+        base: Box::new(AttributeType::String),
+        validate: |value| {
+            if let Value::String(s) = value {
+                validate_prefixed_resource_id(s, "nat")
+                    .map_err(|reason| format!("Invalid NAT Gateway ID '{}': {}", s, reason))
+            } else {
+                Err("Expected string".to_string())
+            }
+        },
+        namespace: None,
+        to_dsl: None,
+    }
+}
+
+/// VPC Peering Connection ID type (e.g., "pcx-12345678")
+pub fn vpc_peering_connection_id() -> AttributeType {
+    AttributeType::Custom {
+        name: "VpcPeeringConnectionId".to_string(),
+        base: Box::new(AttributeType::String),
+        validate: |value| {
+            if let Value::String(s) = value {
+                validate_prefixed_resource_id(s, "pcx").map_err(|reason| {
+                    format!("Invalid VPC Peering Connection ID '{}': {}", s, reason)
+                })
+            } else {
+                Err("Expected string".to_string())
+            }
+        },
+        namespace: None,
+        to_dsl: None,
+    }
+}
+
+/// Transit Gateway ID type (e.g., "tgw-12345678")
+pub fn transit_gateway_id() -> AttributeType {
+    AttributeType::Custom {
+        name: "TransitGatewayId".to_string(),
+        base: Box::new(AttributeType::String),
+        validate: |value| {
+            if let Value::String(s) = value {
+                validate_prefixed_resource_id(s, "tgw")
+                    .map_err(|reason| format!("Invalid Transit Gateway ID '{}': {}", s, reason))
+            } else {
+                Err("Expected string".to_string())
+            }
+        },
+        namespace: None,
+        to_dsl: None,
+    }
+}
+
+/// VPN Gateway ID type (e.g., "vgw-12345678")
+pub fn vpn_gateway_id() -> AttributeType {
+    AttributeType::Custom {
+        name: "VpnGatewayId".to_string(),
+        base: Box::new(AttributeType::String),
+        validate: |value| {
+            if let Value::String(s) = value {
+                validate_prefixed_resource_id(s, "vgw")
+                    .map_err(|reason| format!("Invalid VPN Gateway ID '{}': {}", s, reason))
+            } else {
+                Err("Expected string".to_string())
+            }
+        },
+        namespace: None,
+        to_dsl: None,
+    }
+}
+
+/// Gateway ID type — union of InternetGatewayId and VpnGatewayId.
+pub fn gateway_id() -> AttributeType {
+    AttributeType::Union(vec![internet_gateway_id(), vpn_gateway_id()])
+}
+
+/// Egress Only Internet Gateway ID type (e.g., "eigw-12345678")
+pub fn egress_only_internet_gateway_id() -> AttributeType {
+    AttributeType::Custom {
+        name: "EgressOnlyInternetGatewayId".to_string(),
+        base: Box::new(AttributeType::String),
+        validate: |value| {
+            if let Value::String(s) = value {
+                validate_prefixed_resource_id(s, "eigw").map_err(|reason| {
+                    format!(
+                        "Invalid Egress Only Internet Gateway ID '{}': {}",
+                        s, reason
+                    )
+                })
+            } else {
+                Err("Expected string".to_string())
+            }
+        },
+        namespace: None,
+        to_dsl: None,
+    }
+}
+
+/// VPC Endpoint ID type (e.g., "vpce-12345678")
+pub fn vpc_endpoint_id() -> AttributeType {
+    AttributeType::Custom {
+        name: "VpcEndpointId".to_string(),
+        base: Box::new(AttributeType::String),
+        validate: |value| {
+            if let Value::String(s) = value {
+                validate_prefixed_resource_id(s, "vpce")
+                    .map_err(|reason| format!("Invalid VPC Endpoint ID '{}': {}", s, reason))
+            } else {
+                Err("Expected string".to_string())
+            }
+        },
+        namespace: None,
+        to_dsl: None,
+    }
+}
+
+/// Instance ID type (e.g., "i-0123456789abcdef0")
+pub fn instance_id() -> AttributeType {
+    AttributeType::Custom {
+        name: "InstanceId".to_string(),
+        base: Box::new(AttributeType::String),
+        validate: |value| {
+            if let Value::String(s) = value {
+                validate_prefixed_resource_id(s, "i")
+                    .map_err(|reason| format!("Invalid Instance ID '{}': {}", s, reason))
+            } else {
+                Err("Expected string".to_string())
+            }
+        },
+        namespace: None,
+        to_dsl: None,
+    }
+}
+
+/// Network Interface ID type (e.g., "eni-0123456789abcdef0")
+pub fn network_interface_id() -> AttributeType {
+    AttributeType::Custom {
+        name: "NetworkInterfaceId".to_string(),
+        base: Box::new(AttributeType::String),
+        validate: |value| {
+            if let Value::String(s) = value {
+                validate_prefixed_resource_id(s, "eni")
+                    .map_err(|reason| format!("Invalid Network Interface ID '{}': {}", s, reason))
+            } else {
+                Err("Expected string".to_string())
+            }
+        },
+        namespace: None,
+        to_dsl: None,
+    }
+}
+
+/// EIP Allocation ID type (e.g., "eipalloc-0123456789abcdef0")
+#[allow(dead_code)]
+pub fn allocation_id() -> AttributeType {
+    AttributeType::Custom {
+        name: "AllocationId".to_string(),
+        base: Box::new(AttributeType::String),
+        validate: |value| {
+            if let Value::String(s) = value {
+                validate_prefixed_resource_id(s, "eipalloc")
+                    .map_err(|reason| format!("Invalid Allocation ID '{}': {}", s, reason))
+            } else {
+                Err("Expected string".to_string())
+            }
+        },
+        namespace: None,
+        to_dsl: None,
+    }
+}
+
+/// Prefix List ID type (e.g., "pl-0123456789abcdef0")
+pub fn prefix_list_id() -> AttributeType {
+    AttributeType::Custom {
+        name: "PrefixListId".to_string(),
+        base: Box::new(AttributeType::String),
+        validate: |value| {
+            if let Value::String(s) = value {
+                validate_prefixed_resource_id(s, "pl")
+                    .map_err(|reason| format!("Invalid Prefix List ID '{}': {}", s, reason))
+            } else {
+                Err("Expected string".to_string())
+            }
+        },
+        namespace: None,
+        to_dsl: None,
+    }
+}
+
+/// Carrier Gateway ID type (e.g., "cagw-0123456789abcdef0")
+pub fn carrier_gateway_id() -> AttributeType {
+    AttributeType::Custom {
+        name: "CarrierGatewayId".to_string(),
+        base: Box::new(AttributeType::String),
+        validate: |value| {
+            if let Value::String(s) = value {
+                validate_prefixed_resource_id(s, "cagw")
+                    .map_err(|reason| format!("Invalid Carrier Gateway ID '{}': {}", s, reason))
+            } else {
+                Err("Expected string".to_string())
+            }
+        },
+        namespace: None,
+        to_dsl: None,
+    }
+}
+
+/// Local Gateway ID type (e.g., "lgw-0123456789abcdef0")
+pub fn local_gateway_id() -> AttributeType {
+    AttributeType::Custom {
+        name: "LocalGatewayId".to_string(),
+        base: Box::new(AttributeType::String),
+        validate: |value| {
+            if let Value::String(s) = value {
+                validate_prefixed_resource_id(s, "lgw")
+                    .map_err(|reason| format!("Invalid Local Gateway ID '{}': {}", s, reason))
+            } else {
+                Err("Expected string".to_string())
+            }
+        },
+        namespace: None,
+        to_dsl: None,
+    }
+}
+
+/// Network ACL ID type (e.g., "acl-0123456789abcdef0")
+#[allow(dead_code)]
+pub fn network_acl_id() -> AttributeType {
+    AttributeType::Custom {
+        name: "NetworkAclId".to_string(),
+        base: Box::new(AttributeType::String),
+        validate: |value| {
+            if let Value::String(s) = value {
+                validate_prefixed_resource_id(s, "acl")
+                    .map_err(|reason| format!("Invalid Network ACL ID '{}': {}", s, reason))
+            } else {
+                Err("Expected string".to_string())
+            }
+        },
+        namespace: None,
+        to_dsl: None,
+    }
+}
+
+// ========== AWS Account ID ==========
+
+/// Validate a 12-digit AWS Account ID.
+pub fn validate_aws_account_id(id: &str) -> Result<(), String> {
+    if id.len() != 12 {
+        return Err(format!(
+            "must be exactly 12 digits, got {} characters",
+            id.len()
+        ));
+    }
+    if !id.chars().all(|c| c.is_ascii_digit()) {
+        return Err("must contain only digits".to_string());
+    }
+    Ok(())
+}
+
+/// AWS Account ID type (12-digit numeric string, e.g., "123456789012")
+pub fn aws_account_id() -> AttributeType {
+    AttributeType::Custom {
+        name: "AwsAccountId".to_string(),
+        base: Box::new(AttributeType::String),
+        validate: |value| {
+            if let Value::String(s) = value {
+                validate_aws_account_id(s)
+                    .map_err(|reason| format!("Invalid AWS Account ID '{}': {}", s, reason))
+            } else {
+                Err("Expected string".to_string())
+            }
+        },
+        namespace: None,
+        to_dsl: None,
+    }
+}
+
+// ========== ARN validators ==========
+
+/// Validate basic ARN format (starts with "arn:", has 6+ colon-separated parts).
+pub fn validate_arn(arn: &str) -> Result<(), String> {
+    if !arn.starts_with("arn:") {
+        return Err("must start with 'arn:'".to_string());
+    }
+    let parts: Vec<&str> = arn.splitn(6, ':').collect();
+    if parts.len() < 6 {
+        return Err(
+            "must have at least 6 colon-separated parts (arn:partition:service:region:account:resource)".to_string()
+        );
+    }
+    Ok(())
+}
+
+/// Validate an ARN for a specific AWS service and optional resource prefix.
+pub fn validate_service_arn(
+    arn: &str,
+    expected_service: &str,
+    resource_prefix: Option<&str>,
+) -> Result<(), String> {
+    validate_arn(arn)?;
+    let parts: Vec<&str> = arn.splitn(6, ':').collect();
+    if parts[2] != expected_service {
+        return Err(format!(
+            "expected {} service, got '{}'",
+            expected_service, parts[2]
+        ));
+    }
+    if let Some(prefix) = resource_prefix
+        && !parts[5].starts_with(prefix)
+    {
+        return Err(format!(
+            "expected resource starting with '{}', got '{}'",
+            prefix, parts[5]
+        ));
+    }
+    Ok(())
+}
+
+/// ARN type (e.g., "arn:aws:s3:::my-bucket")
+pub fn arn() -> AttributeType {
+    AttributeType::Custom {
+        name: "Arn".to_string(),
+        base: Box::new(AttributeType::String),
+        validate: |value| {
+            if let Value::String(s) = value {
+                validate_arn(s).map_err(|reason| format!("Invalid ARN '{}': {}", s, reason))
+            } else {
+                Err("Expected string".to_string())
+            }
+        },
+        namespace: None,
+        to_dsl: None,
+    }
+}
+
+/// IAM Role ARN type (e.g., "arn:aws:iam::123456789012:role/MyRole")
+#[allow(dead_code)]
+pub fn iam_role_arn() -> AttributeType {
+    AttributeType::Custom {
+        name: "IamRoleArn".to_string(),
+        base: Box::new(AttributeType::String),
+        validate: |value| {
+            if let Value::String(s) = value {
+                validate_service_arn(s, "iam", Some("role/"))
+                    .map_err(|reason| format!("Invalid IAM Role ARN '{}': {}", s, reason))
+            } else {
+                Err("Expected string".to_string())
+            }
+        },
+        namespace: None,
+        to_dsl: None,
+    }
+}
+
+/// IAM Policy ARN type (e.g., "arn:aws:iam::123456789012:policy/MyPolicy")
+#[allow(dead_code)]
+pub fn iam_policy_arn() -> AttributeType {
+    AttributeType::Custom {
+        name: "IamPolicyArn".to_string(),
+        base: Box::new(AttributeType::String),
+        validate: |value| {
+            if let Value::String(s) = value {
+                validate_service_arn(s, "iam", Some("policy/"))
+                    .map_err(|reason| format!("Invalid IAM Policy ARN '{}': {}", s, reason))
+            } else {
+                Err("Expected string".to_string())
+            }
+        },
+        namespace: None,
+        to_dsl: None,
+    }
+}
+
+/// KMS Key ARN type (e.g., "arn:aws:kms:us-east-1:123456789012:key/...")
+#[allow(dead_code)]
+pub fn kms_key_arn() -> AttributeType {
+    AttributeType::Custom {
+        name: "KmsKeyArn".to_string(),
+        base: Box::new(AttributeType::String),
+        validate: |value| {
+            if let Value::String(s) = value {
+                validate_service_arn(s, "kms", Some("key/"))
+                    .map_err(|reason| format!("Invalid KMS Key ARN '{}': {}", s, reason))
+            } else {
+                Err("Expected string".to_string())
+            }
+        },
+        namespace: None,
+        to_dsl: None,
+    }
+}
+
+// ========== KMS Key ID ==========
+
+/// Check if a string is a valid UUID (8-4-4-4-12 hex digits).
+pub fn is_uuid(s: &str) -> bool {
+    let expected_lens = [8, 4, 4, 4, 12];
+    let parts: Vec<&str> = s.split('-').collect();
+    parts.len() == 5
+        && parts
+            .iter()
+            .zip(expected_lens.iter())
+            .all(|(part, &len)| part.len() == len && part.chars().all(|c| c.is_ascii_hexdigit()))
+}
+
+/// Validate a KMS Key ID (ARN, alias, or UUID format).
+pub fn validate_kms_key_id(value: &str) -> Result<(), String> {
+    // Accept KMS ARNs with key/ or alias/ resource prefix
+    if value.starts_with("arn:") {
+        validate_service_arn(value, "kms", None)?;
+        let parts: Vec<&str> = value.splitn(6, ':').collect();
+        let resource = parts[5];
+        if !resource.starts_with("key/") && !resource.starts_with("alias/") {
+            return Err(format!(
+                "KMS ARN resource '{}' must start with 'key/' or 'alias/'",
+                resource
+            ));
+        }
+        return Ok(());
+    }
+    // Accept alias format: alias/<name>
+    if value.starts_with("alias/") {
+        if value.len() <= "alias/".len() {
+            return Err("missing alias name after 'alias/'".to_string());
+        }
+        return Ok(());
+    }
+    // Accept bare key ID (UUID format: 8-4-4-4-12 hex digits)
+    if is_uuid(value) {
+        return Ok(());
+    }
+    Err(
+        "expected a key ARN, alias ARN, alias name (alias/...), or key ID (UUID format)"
+            .to_string(),
+    )
+}
+
+/// KMS Key ID type - accepts multiple formats:
+/// - Key ARN: "arn:aws:kms:us-east-1:123456789012:key/..."
+/// - Key alias ARN: "arn:aws:kms:us-east-1:123456789012:alias/my-key"
+/// - Key alias: "alias/my-key"
+/// - Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"
+#[allow(dead_code)]
+pub fn kms_key_id() -> AttributeType {
+    AttributeType::Custom {
+        name: "KmsKeyId".to_string(),
+        base: Box::new(AttributeType::String),
+        validate: |value| {
+            if let Value::String(s) = value {
+                validate_kms_key_id(s)
+                    .map_err(|reason| format!("Invalid KMS key identifier '{}': {}", s, reason))
+            } else {
+                Err("Expected string".to_string())
+            }
+        },
+        namespace: None,
+        to_dsl: None,
+    }
+}
+
+// ========== IPAM types ==========
+
+/// Validate IPAM Pool ID format: `ipam-pool-{hex}` where hex is 8+ hex digits.
+pub fn validate_ipam_pool_id(id: &str) -> Result<(), String> {
+    let Some(hex_part) = id.strip_prefix("ipam-pool-") else {
+        return Err("expected format 'ipam-pool-{hex}'".to_string());
+    };
+    if hex_part.len() < 8 {
+        return Err("hex part must be at least 8 characters".to_string());
+    }
+    if !hex_part.chars().all(|c| c.is_ascii_hexdigit()) {
+        return Err("hex part must contain only hex digits".to_string());
+    }
+    Ok(())
+}
+
+/// IPAM Pool ID type (e.g., "ipam-pool-0123456789abcdef0")
+pub fn ipam_pool_id() -> AttributeType {
+    AttributeType::Custom {
+        name: "IpamPoolId".to_string(),
+        base: Box::new(AttributeType::String),
+        validate: |value| {
+            if let Value::String(s) = value {
+                validate_ipam_pool_id(s)
+                    .map_err(|reason| format!("Invalid IPAM Pool ID '{}': {}", s, reason))
+            } else {
+                Err("Expected string".to_string())
+            }
+        },
+        namespace: None,
+        to_dsl: None,
+    }
+}
+
+// ========== Availability Zone ==========
+
+/// Validate availability zone format (e.g., "us-east-1a").
+pub fn validate_availability_zone(az: &str) -> Result<(), String> {
+    // Must end with a single lowercase letter (zone identifier)
+    let zone_letter = az.chars().last();
+    if !zone_letter.is_some_and(|c| c.is_ascii_lowercase()) {
+        return Err("must end with a zone letter (a-z)".to_string());
+    }
+
+    // Region part is everything except the last character
+    let region = &az[..az.len() - 1];
+
+    // Region must match pattern: lowercase-lowercase-digit
+    // e.g., "us-east-1", "ap-northeast-1", "eu-west-2"
+    let parts: Vec<&str> = region.split('-').collect();
+    if parts.len() < 3 {
+        return Err("expected format like 'us-east-1a'".to_string());
+    }
+
+    // Last part of region must be a number
+    let last = parts.last().unwrap();
+    if last.parse::<u8>().is_err() {
+        return Err("region must end with a number".to_string());
+    }
+
+    // All other parts must be lowercase alphabetic
+    for part in &parts[..parts.len() - 1] {
+        if part.is_empty() || !part.chars().all(|c| c.is_ascii_lowercase()) {
+            return Err("expected format like 'us-east-1a'".to_string());
+        }
+    }
+
+    Ok(())
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    // ARN tests
+
+    #[test]
+    fn validate_arn_valid() {
+        assert!(validate_arn("arn:aws:s3:::my-bucket").is_ok());
+        assert!(validate_arn("arn:aws:iam::123456789012:role/MyRole").is_ok());
+        assert!(validate_arn("arn:aws-cn:s3:::my-bucket").is_ok());
+        assert!(validate_arn("arn:aws:ec2:us-east-1:123456789012:vpc/vpc-1234").is_ok());
+    }
+
+    #[test]
+    fn validate_arn_invalid() {
+        assert!(validate_arn("not-an-arn").is_err());
+        assert!(validate_arn("arn:aws:s3").is_err());
+        assert!(validate_arn("arn:aws").is_err());
+        assert!(validate_arn("").is_err());
+    }
+
+    #[test]
+    fn validate_arn_type_with_value() {
+        let t = arn();
+        assert!(
+            t.validate(&Value::String("arn:aws:s3:::my-bucket".to_string()))
+                .is_ok()
+        );
+        assert!(
+            t.validate(&Value::String("not-an-arn".to_string()))
+                .is_err()
+        );
+        assert!(t.validate(&Value::Int(42)).is_err());
+        assert!(
+            t.validate(&Value::ResourceRef {
+                binding_name: "role".to_string(),
+                attribute_name: "arn".to_string(),
+            })
+            .is_ok()
+        );
+    }
+
+    // Resource ID tests
+
+    #[test]
+    fn validate_aws_resource_id_valid() {
+        assert!(validate_aws_resource_id("vpc-1a2b3c4d").is_ok());
+        assert!(validate_aws_resource_id("subnet-0123456789abcdef0").is_ok());
+        assert!(validate_aws_resource_id("sg-12345678").is_ok());
+        assert!(validate_aws_resource_id("rtb-abcdef12").is_ok());
+        assert!(validate_aws_resource_id("eipalloc-0123456789abcdef0").is_ok());
+        assert!(validate_aws_resource_id("igw-12345678").is_ok());
+    }
+
+    #[test]
+    fn validate_aws_resource_id_invalid() {
+        assert!(validate_aws_resource_id("not-a-valid-id").is_err());
+        assert!(validate_aws_resource_id("vpc").is_err());
+        assert!(validate_aws_resource_id("vpc-short").is_err());
+        assert!(validate_aws_resource_id("vpc-1234567").is_err());
+        assert!(validate_aws_resource_id("VPC-12345678").is_err());
+    }
+
+    #[test]
+    fn validate_aws_resource_id_type_with_value() {
+        let t = aws_resource_id();
+        assert!(
+            t.validate(&Value::String("vpc-1a2b3c4d".to_string()))
+                .is_ok()
+        );
+        assert!(t.validate(&Value::String("vpc".to_string())).is_err());
+        assert!(t.validate(&Value::Int(42)).is_err());
+        assert!(
+            t.validate(&Value::ResourceRef {
+                binding_name: "my_vpc".to_string(),
+                attribute_name: "vpc_id".to_string(),
+            })
+            .is_ok()
+        );
+    }
+
+    // Availability zone tests
+
+    #[test]
+    fn validate_availability_zone_valid() {
+        assert!(validate_availability_zone("us-east-1a").is_ok());
+        assert!(validate_availability_zone("ap-northeast-1c").is_ok());
+        assert!(validate_availability_zone("eu-central-1b").is_ok());
+        assert!(validate_availability_zone("me-south-1a").is_ok());
+        assert!(validate_availability_zone("us-west-2d").is_ok());
+    }
+
+    #[test]
+    fn validate_availability_zone_invalid() {
+        assert!(validate_availability_zone("us-east-1").is_err()); // no zone letter
+        assert!(validate_availability_zone("a").is_err()); // too short
+        assert!(validate_availability_zone("invalid").is_err());
+    }
+
+    // Enum helpers
+
+    #[test]
+    fn find_matching_enum_value_exact() {
+        let values = &["Enabled", "Suspended"];
+        assert_eq!(find_matching_enum_value("Enabled", values), Some("Enabled"));
+        assert_eq!(find_matching_enum_value("Missing", values), None);
+    }
+
+    #[test]
+    fn find_matching_enum_value_case_insensitive() {
+        let values = &["Enabled", "Suspended"];
+        assert_eq!(find_matching_enum_value("enabled", values), Some("Enabled"));
+    }
+
+    #[test]
+    fn find_matching_enum_value_underscore_to_hyphen() {
+        let values = &["us-east-1", "eu-west-1"];
+        assert_eq!(
+            find_matching_enum_value("us_east_1", values),
+            Some("us-east-1")
+        );
+    }
+
+    #[test]
+    fn canonicalize_enum_value_matches() {
+        assert_eq!(
+            canonicalize_enum_value("enabled", &["Enabled", "Suspended"]),
+            "Enabled"
+        );
+    }
+
+    #[test]
+    fn canonicalize_enum_value_no_match() {
+        assert_eq!(
+            canonicalize_enum_value("unknown", &["Enabled", "Suspended"]),
+            "unknown"
+        );
+    }
+
+    // IPAM Pool ID tests
+
+    #[test]
+    fn validate_ipam_pool_id_valid() {
+        assert!(validate_ipam_pool_id("ipam-pool-0123456789abcdef0").is_ok());
+        assert!(validate_ipam_pool_id("ipam-pool-12345678").is_ok());
+    }
+
+    #[test]
+    fn validate_ipam_pool_id_invalid() {
+        assert!(validate_ipam_pool_id("ipam-pool-short").is_err());
+        assert!(validate_ipam_pool_id("not-ipam-pool").is_err());
+        assert!(validate_ipam_pool_id("ipam-pool-").is_err());
+    }
+
+    // AWS Account ID tests
+
+    #[test]
+    fn validate_aws_account_id_valid() {
+        assert!(validate_aws_account_id("123456789012").is_ok());
+    }
+
+    #[test]
+    fn validate_aws_account_id_invalid() {
+        assert!(validate_aws_account_id("1234").is_err());
+        assert!(validate_aws_account_id("12345678901a").is_err());
+        assert!(validate_aws_account_id("").is_err());
+    }
+
+    // KMS Key ID tests
+
+    #[test]
+    fn validate_kms_key_id_arn() {
+        assert!(
+            validate_kms_key_id(
+                "arn:aws:kms:us-east-1:123456789012:key/1234abcd-12ab-34cd-56ef-1234567890ab"
+            )
+            .is_ok()
+        );
+        assert!(validate_kms_key_id("arn:aws:kms:us-east-1:123456789012:alias/my-key").is_ok());
+    }
+
+    #[test]
+    fn validate_kms_key_id_alias() {
+        assert!(validate_kms_key_id("alias/my-key").is_ok());
+        assert!(validate_kms_key_id("alias/").is_err());
+    }
+
+    #[test]
+    fn validate_kms_key_id_uuid() {
+        assert!(validate_kms_key_id("1234abcd-12ab-34cd-56ef-1234567890ab").is_ok());
+        assert!(validate_kms_key_id("not-a-uuid").is_err());
+    }
+
+    // Service ARN tests
+
+    #[test]
+    fn validate_service_arn_valid() {
+        assert!(
+            validate_service_arn(
+                "arn:aws:iam::123456789012:role/MyRole",
+                "iam",
+                Some("role/")
+            )
+            .is_ok()
+        );
+    }
+
+    #[test]
+    fn validate_service_arn_wrong_service() {
+        assert!(validate_service_arn("arn:aws:s3:::bucket", "iam", None).is_err());
+    }
+
+    #[test]
+    fn validate_service_arn_wrong_prefix() {
+        assert!(
+            validate_service_arn(
+                "arn:aws:iam::123456789012:user/MyUser",
+                "iam",
+                Some("role/")
+            )
+            .is_err()
+        );
+    }
+
+    // UUID tests
+
+    #[test]
+    fn is_uuid_valid() {
+        assert!(is_uuid("1234abcd-12ab-34cd-56ef-1234567890ab"));
+    }
+
+    #[test]
+    fn is_uuid_invalid() {
+        assert!(!is_uuid("not-a-uuid"));
+        assert!(!is_uuid("1234abcd-12ab-34cd-56ef"));
+        assert!(!is_uuid(""));
+    }
+}
