+++ carina-provider-aws/Cargo.toml
@@ -9,6 +9,7 @@ doctest = false
 
 [dependencies]
 carina-core = { path = "../carina-core" }
+carina-aws-types = { path = "../carina-aws-types" }
 aws-config = "1"
 aws-sdk-s3 = "1"
 aws-sdk-ec2 = "1"
