+++ carina-provider-aws/src/schemas/types.rs
@@ -1,9 +1,9 @@
 //! AWS-specific type definitions and validators
 //!
-//! This module contains stable utility code for AWS-specific types,
-//! validators, and enum normalization. These functions are NOT generated
-//! from CloudFormation schemas — they are hand-written and imported by
-//! the generated `mod.rs`.
+//! This module re-exports shared AWS type validators from `carina-aws-types`
+//! and defines provider-specific types (region, availability zone, schema config).
+
+pub use carina_aws_types::*;
 
 use carina_core::resource::Value;
 use carina_core::schema::{AttributeType, ResourceSchema};
@@ -24,67 +24,6 @@ pub struct AwsSchemaConfig {
     /// The resource schema with attribute definitions
     pub schema: ResourceSchema,
 }
-
-/// Tags type for AWS resources (Terraform-style map)
-pub(crate) fn tags_type() -> AttributeType {
-    AttributeType::Map(Box::new(AttributeType::String))
-}
-
-/// Check if `input` matches any of `valid_values` using enum matching rules:
-/// exact match, case-insensitive, or underscore-to-hyphen (case-insensitive).
-/// Returns the matched valid value if found.
-fn find_matching_enum_value<'a>(input: &str, valid_values: &[&'a str]) -> Option<&'a str> {
-    // Exact match
-    if let Some(&v) = valid_values.iter().find(|&&v| v == input) {
-        return Some(v);
-    }
-    // Case-insensitive match
-    if let Some(&v) = valid_values
-        .iter()
-        .find(|&&v| v.eq_ignore_ascii_case(input))
-    {
-        return Some(v);
-    }
-    // Underscore-to-hyphen match (case-insensitive)
-    let hyphenated = input.replace('_', "-");
-    if let Some(&v) = valid_values
-        .iter()
-        .find(|&&v| v.eq_ignore_ascii_case(&hyphenated))
-    {
-        return Some(v);
-    }
-    None
-}
-
-/// Canonicalize an enum value by matching against valid values.
-/// Handles exact match, case-insensitive match, and underscore-to-hyphen conversion.
-#[allow(dead_code)]
-pub(crate) fn canonicalize_enum_value(raw: &str, valid_values: &[&str]) -> String {
-    find_matching_enum_value(raw, valid_values)
-        .unwrap_or(raw)
-        .to_string()
-}
-
-/// Valid AWS regions (in AWS format with hyphens)
-const VALID_REGIONS: &[&str] = &[
-    "ap-northeast-1",
-    "ap-northeast-2",
-    "ap-northeast-3",
-    "ap-southeast-1",
-    "ap-southeast-2",
-    "ap-south-1",
-    "us-east-1",
-    "us-east-2",
-    "us-west-1",
-    "us-west-2",
-    "eu-west-1",
-    "eu-west-2",
-    "eu-west-3",
-    "eu-central-1",
-    "eu-north-1",
-    "ca-central-1",
-    "sa-east-1",
-];
 
 /// AWS region type with custom validation
 /// Accepts:
@@ -119,595 +58,16 @@ pub fn aws_region() -> AttributeType {
     }
 }
 
-// ========== Resource ID validators ==========
-
-/// Validate a resource ID with a specific prefix
-fn validate_prefixed_resource_id(id: &str, expected_prefix: &str) -> Result<(), String> {
-    let expected_format = format!("{}-xxxxxxxx", expected_prefix);
-    if !id.starts_with(&format!("{}-", expected_prefix)) {
-        return Err(format!("expected format '{}'", expected_format));
-    }
-    validate_aws_resource_id(id)
-}
-
-fn validate_aws_resource_id(id: &str) -> Result<(), String> {
-    let Some(dash_pos) = id.find('-') else {
-        return Err("expected format 'prefix-hexdigits'".to_string());
-    };
-
-    let prefix = &id[..dash_pos];
-    let hex_part = &id[dash_pos + 1..];
-
-    if prefix.is_empty()
-        || !prefix
-            .chars()
-            .all(|c| c.is_ascii_lowercase() || c.is_ascii_digit())
-    {
-        return Err("prefix must be lowercase alphanumeric".to_string());
-    }
-
-    if hex_part.len() < 8 {
-        return Err("ID part must be at least 8 characters after prefix".to_string());
-    }
-
-    if !hex_part.chars().all(|c| c.is_ascii_hexdigit()) {
-        return Err("ID part must contain only hex digits".to_string());
-    }
-
-    Ok(())
-}
-
-/// AWS resource ID type (e.g., "vpc-1a2b3c4d", "subnet-0123456789abcdef0")
-#[allow(dead_code)]
-pub(crate) fn aws_resource_id() -> AttributeType {
-    AttributeType::Custom {
-        name: "AwsResourceId".to_string(),
-        base: Box::new(AttributeType::String),
-        validate: |value| {
-            if let Value::String(s) = value {
-                validate_aws_resource_id(s)
-                    .map_err(|reason| format!("Invalid resource ID '{}': {}", s, reason))
-            } else {
-                Err("Expected string".to_string())
-            }
-        },
-        namespace: None,
-        to_dsl: None,
-    }
-}
-
-/// VPC ID type (e.g., "vpc-1a2b3c4d")
-pub(crate) fn vpc_id() -> AttributeType {
-    AttributeType::Custom {
-        name: "VpcId".to_string(),
-        base: Box::new(AttributeType::String),
-        validate: |value| {
-            if let Value::String(s) = value {
-                validate_prefixed_resource_id(s, "vpc")
-                    .map_err(|reason| format!("Invalid VPC ID '{}': {}", s, reason))
-            } else {
-                Err("Expected string".to_string())
-            }
-        },
-        namespace: None,
-        to_dsl: None,
-    }
-}
-
-/// Subnet ID type (e.g., "subnet-0123456789abcdef0")
-pub(crate) fn subnet_id() -> AttributeType {
-    AttributeType::Custom {
-        name: "SubnetId".to_string(),
-        base: Box::new(AttributeType::String),
-        validate: |value| {
-            if let Value::String(s) = value {
-                validate_prefixed_resource_id(s, "subnet")
-                    .map_err(|reason| format!("Invalid Subnet ID '{}': {}", s, reason))
-            } else {
-                Err("Expected string".to_string())
-            }
-        },
-        namespace: None,
-        to_dsl: None,
-    }
-}
-
-/// Security Group ID type (e.g., "sg-12345678")
-pub(crate) fn security_group_id() -> AttributeType {
-    AttributeType::Custom {
-        name: "SecurityGroupId".to_string(),
-        base: Box::new(AttributeType::String),
-        validate: |value| {
-            if let Value::String(s) = value {
-                validate_prefixed_resource_id(s, "sg")
-                    .map_err(|reason| format!("Invalid Security Group ID '{}': {}", s, reason))
-            } else {
-                Err("Expected string".to_string())
-            }
-        },
-        namespace: None,
-        to_dsl: None,
-    }
-}
-
-/// Internet Gateway ID type (e.g., "igw-12345678")
-pub(crate) fn internet_gateway_id() -> AttributeType {
-    AttributeType::Custom {
-        name: "InternetGatewayId".to_string(),
-        base: Box::new(AttributeType::String),
-        validate: |value| {
-            if let Value::String(s) = value {
-                validate_prefixed_resource_id(s, "igw")
-                    .map_err(|reason| format!("Invalid Internet Gateway ID '{}': {}", s, reason))
-            } else {
-                Err("Expected string".to_string())
-            }
-        },
-        namespace: None,
-        to_dsl: None,
-    }
-}
-
-/// Route Table ID type (e.g., "rtb-abcdef12")
-pub(crate) fn route_table_id() -> AttributeType {
-    AttributeType::Custom {
-        name: "RouteTableId".to_string(),
-        base: Box::new(AttributeType::String),
-        validate: |value| {
-            if let Value::String(s) = value {
-                validate_prefixed_resource_id(s, "rtb")
-                    .map_err(|reason| format!("Invalid Route Table ID '{}': {}", s, reason))
-            } else {
-                Err("Expected string".to_string())
-            }
-        },
-        namespace: None,
-        to_dsl: None,
-    }
-}
-
-/// NAT Gateway ID type (e.g., "nat-12345678")
-pub(crate) fn nat_gateway_id() -> AttributeType {
-    AttributeType::Custom {
-        name: "NatGatewayId".to_string(),
-        base: Box::new(AttributeType::String),
-        validate: |value| {
-            if let Value::String(s) = value {
-                validate_prefixed_resource_id(s, "nat")
-                    .map_err(|reason| format!("Invalid NAT Gateway ID '{}': {}", s, reason))
-            } else {
-                Err("Expected string".to_string())
-            }
-        },
-        namespace: None,
-        to_dsl: None,
-    }
-}
-
-/// VPC Peering Connection ID type (e.g., "pcx-12345678")
-pub(crate) fn vpc_peering_connection_id() -> AttributeType {
-    AttributeType::Custom {
-        name: "VpcPeeringConnectionId".to_string(),
-        base: Box::new(AttributeType::String),
-        validate: |value| {
-            if let Value::String(s) = value {
-                validate_prefixed_resource_id(s, "pcx").map_err(|reason| {
-                    format!("Invalid VPC Peering Connection ID '{}': {}", s, reason)
-                })
-            } else {
-                Err("Expected string".to_string())
-            }
-        },
-        namespace: None,
-        to_dsl: None,
-    }
-}
-
-/// Transit Gateway ID type (e.g., "tgw-12345678")
-pub(crate) fn transit_gateway_id() -> AttributeType {
-    AttributeType::Custom {
-        name: "TransitGatewayId".to_string(),
-        base: Box::new(AttributeType::String),
-        validate: |value| {
-            if let Value::String(s) = value {
-                validate_prefixed_resource_id(s, "tgw")
-                    .map_err(|reason| format!("Invalid Transit Gateway ID '{}': {}", s, reason))
-            } else {
-                Err("Expected string".to_string())
-            }
-        },
-        namespace: None,
-        to_dsl: None,
-    }
-}
-
-/// VPN Gateway ID type (e.g., "vgw-12345678")
-pub(crate) fn vpn_gateway_id() -> AttributeType {
-    AttributeType::Custom {
-        name: "VpnGatewayId".to_string(),
-        base: Box::new(AttributeType::String),
-        validate: |value| {
-            if let Value::String(s) = value {
-                validate_prefixed_resource_id(s, "vgw")
-                    .map_err(|reason| format!("Invalid VPN Gateway ID '{}': {}", s, reason))
-            } else {
-                Err("Expected string".to_string())
-            }
-        },
-        namespace: None,
-        to_dsl: None,
-    }
-}
-
-/// Egress Only Internet Gateway ID type (e.g., "eigw-12345678")
-pub(crate) fn egress_only_internet_gateway_id() -> AttributeType {
-    AttributeType::Custom {
-        name: "EgressOnlyInternetGatewayId".to_string(),
-        base: Box::new(AttributeType::String),
-        validate: |value| {
-            if let Value::String(s) = value {
-                validate_prefixed_resource_id(s, "eigw").map_err(|reason| {
-                    format!(
-                        "Invalid Egress Only Internet Gateway ID '{}': {}",
-                        s, reason
-                    )
-                })
-            } else {
-                Err("Expected string".to_string())
-            }
-        },
-        namespace: None,
-        to_dsl: None,
-    }
-}
-
-/// VPC Endpoint ID type (e.g., "vpce-12345678")
-pub(crate) fn vpc_endpoint_id() -> AttributeType {
-    AttributeType::Custom {
-        name: "VpcEndpointId".to_string(),
-        base: Box::new(AttributeType::String),
-        validate: |value| {
-            if let Value::String(s) = value {
-                validate_prefixed_resource_id(s, "vpce")
-                    .map_err(|reason| format!("Invalid VPC Endpoint ID '{}': {}", s, reason))
-            } else {
-                Err("Expected string".to_string())
-            }
-        },
-        namespace: None,
-        to_dsl: None,
-    }
-}
-
-/// Instance ID type (e.g., "i-0123456789abcdef0")
-pub(crate) fn instance_id() -> AttributeType {
-    AttributeType::Custom {
-        name: "InstanceId".to_string(),
-        base: Box::new(AttributeType::String),
-        validate: |value| {
-            if let Value::String(s) = value {
-                validate_prefixed_resource_id(s, "i")
-                    .map_err(|reason| format!("Invalid Instance ID '{}': {}", s, reason))
-            } else {
-                Err("Expected string".to_string())
-            }
-        },
-        namespace: None,
-        to_dsl: None,
-    }
-}
-
-/// Network Interface ID type (e.g., "eni-0123456789abcdef0")
-pub(crate) fn network_interface_id() -> AttributeType {
-    AttributeType::Custom {
-        name: "NetworkInterfaceId".to_string(),
-        base: Box::new(AttributeType::String),
-        validate: |value| {
-            if let Value::String(s) = value {
-                validate_prefixed_resource_id(s, "eni")
-                    .map_err(|reason| format!("Invalid Network Interface ID '{}': {}", s, reason))
-            } else {
-                Err("Expected string".to_string())
-            }
-        },
-        namespace: None,
-        to_dsl: None,
-    }
-}
-
-/// EIP Allocation ID type (e.g., "eipalloc-0123456789abcdef0")
-#[allow(dead_code)]
-pub(crate) fn allocation_id() -> AttributeType {
-    AttributeType::Custom {
-        name: "AllocationId".to_string(),
-        base: Box::new(AttributeType::String),
-        validate: |value| {
-            if let Value::String(s) = value {
-                validate_prefixed_resource_id(s, "eipalloc")
-                    .map_err(|reason| format!("Invalid Allocation ID '{}': {}", s, reason))
-            } else {
-                Err("Expected string".to_string())
-            }
-        },
-        namespace: None,
-        to_dsl: None,
-    }
-}
-
-/// Prefix List ID type (e.g., "pl-0123456789abcdef0")
-pub(crate) fn prefix_list_id() -> AttributeType {
-    AttributeType::Custom {
-        name: "PrefixListId".to_string(),
-        base: Box::new(AttributeType::String),
-        validate: |value| {
-            if let Value::String(s) = value {
-                validate_prefixed_resource_id(s, "pl")
-                    .map_err(|reason| format!("Invalid Prefix List ID '{}': {}", s, reason))
-            } else {
-                Err("Expected string".to_string())
-            }
-        },
-        namespace: None,
-        to_dsl: None,
-    }
-}
-
-/// Carrier Gateway ID type (e.g., "cagw-0123456789abcdef0")
-pub(crate) fn carrier_gateway_id() -> AttributeType {
-    AttributeType::Custom {
-        name: "CarrierGatewayId".to_string(),
-        base: Box::new(AttributeType::String),
-        validate: |value| {
-            if let Value::String(s) = value {
-                validate_prefixed_resource_id(s, "cagw")
-                    .map_err(|reason| format!("Invalid Carrier Gateway ID '{}': {}", s, reason))
-            } else {
-                Err("Expected string".to_string())
-            }
-        },
-        namespace: None,
-        to_dsl: None,
-    }
-}
-
-/// Local Gateway ID type (e.g., "lgw-0123456789abcdef0")
-pub(crate) fn local_gateway_id() -> AttributeType {
-    AttributeType::Custom {
-        name: "LocalGatewayId".to_string(),
-        base: Box::new(AttributeType::String),
-        validate: |value| {
-            if let Value::String(s) = value {
-                validate_prefixed_resource_id(s, "lgw")
-                    .map_err(|reason| format!("Invalid Local Gateway ID '{}': {}", s, reason))
-            } else {
-                Err("Expected string".to_string())
-            }
-        },
-        namespace: None,
-        to_dsl: None,
-    }
-}
-
-/// Network ACL ID type (e.g., "acl-0123456789abcdef0")
-#[allow(dead_code)]
-pub(crate) fn network_acl_id() -> AttributeType {
-    AttributeType::Custom {
-        name: "NetworkAclId".to_string(),
-        base: Box::new(AttributeType::String),
-        validate: |value| {
-            if let Value::String(s) = value {
-                validate_prefixed_resource_id(s, "acl")
-                    .map_err(|reason| format!("Invalid Network ACL ID '{}': {}", s, reason))
-            } else {
-                Err("Expected string".to_string())
-            }
-        },
-        namespace: None,
-        to_dsl: None,
-    }
-}
-
-/// AWS Account ID type (12-digit numeric string, e.g., "123456789012")
-pub(crate) fn aws_account_id() -> AttributeType {
-    AttributeType::Custom {
-        name: "AwsAccountId".to_string(),
-        base: Box::new(AttributeType::String),
-        validate: |value| {
-            if let Value::String(s) = value {
-                validate_aws_account_id(s)
-                    .map_err(|reason| format!("Invalid AWS Account ID '{}': {}", s, reason))
-            } else {
-                Err("Expected string".to_string())
-            }
-        },
-        namespace: None,
-        to_dsl: None,
-    }
-}
-
-fn validate_aws_account_id(id: &str) -> Result<(), String> {
-    if id.len() != 12 {
-        return Err(format!(
-            "must be exactly 12 digits, got {} characters",
-            id.len()
-        ));
-    }
-    if !id.chars().all(|c| c.is_ascii_digit()) {
-        return Err("must contain only digits".to_string());
-    }
-    Ok(())
-}
-
-/// Gateway ID type — union of InternetGatewayId and VpnGatewayId.
-/// Used for attributes like ec2_route.gateway_id that accept both igw-* and vgw-* IDs.
-pub(crate) fn gateway_id() -> AttributeType {
-    AttributeType::Union(vec![internet_gateway_id(), vpn_gateway_id()])
-}
-
-// ========== ARN validators ==========
-
-/// ARN type (e.g., "arn:aws:s3:::my-bucket")
-pub(crate) fn arn() -> AttributeType {
-    AttributeType::Custom {
-        name: "Arn".to_string(),
-        base: Box::new(AttributeType::String),
-        validate: |value| {
-            if let Value::String(s) = value {
-                validate_arn(s).map_err(|reason| format!("Invalid ARN '{}': {}", s, reason))
-            } else {
-                Err("Expected string".to_string())
-            }
-        },
-        namespace: None,
-        to_dsl: None,
-    }
-}
-
-pub fn validate_arn(arn: &str) -> Result<(), String> {
-    if !arn.starts_with("arn:") {
-        return Err("must start with 'arn:'".to_string());
-    }
-    let parts: Vec<&str> = arn.splitn(6, ':').collect();
-    if parts.len() < 6 {
-        return Err(
-            "must have at least 6 colon-separated parts (arn:partition:service:region:account:resource)".to_string()
-        );
-    }
-    Ok(())
-}
-
-/// IAM Role ARN type
-#[allow(dead_code)]
-pub(crate) fn iam_role_arn() -> AttributeType {
-    AttributeType::Custom {
-        name: "IamRoleArn".to_string(),
-        base: Box::new(AttributeType::String),
-        validate: |value| {
-            if let Value::String(s) = value {
-                validate_arn(s)
-                    .map_err(|reason| format!("Invalid IAM Role ARN '{}': {}", s, reason))
-            } else {
-                Err("Expected string".to_string())
-            }
-        },
-        namespace: None,
-        to_dsl: None,
-    }
-}
-
-/// IAM Policy ARN type
-#[allow(dead_code)]
-pub(crate) fn iam_policy_arn() -> AttributeType {
-    AttributeType::Custom {
-        name: "IamPolicyArn".to_string(),
-        base: Box::new(AttributeType::String),
-        validate: |value| {
-            if let Value::String(s) = value {
-                validate_arn(s)
-                    .map_err(|reason| format!("Invalid IAM Policy ARN '{}': {}", s, reason))
-            } else {
-                Err("Expected string".to_string())
-            }
-        },
-        namespace: None,
-        to_dsl: None,
-    }
-}
-
-/// KMS Key ARN type
-#[allow(dead_code)]
-pub(crate) fn kms_key_arn() -> AttributeType {
-    AttributeType::Custom {
-        name: "KmsKeyArn".to_string(),
-        base: Box::new(AttributeType::String),
-        validate: |value| {
-            if let Value::String(s) = value {
-                validate_arn(s).map_err(|reason| format!("Invalid KMS Key ARN '{}': {}", s, reason))
-            } else {
-                Err("Expected string".to_string())
-            }
-        },
-        namespace: None,
-        to_dsl: None,
-    }
-}
-
-/// KMS Key ID type (can be ARN, alias, or UUID)
-#[allow(dead_code)]
-pub(crate) fn kms_key_id() -> AttributeType {
-    AttributeType::Custom {
-        name: "KmsKeyId".to_string(),
-        base: Box::new(AttributeType::String),
-        validate: |value| {
-            if let Value::String(_s) = value {
-                Ok(()) // KMS key ID accepts many formats
-            } else {
-                Err("Expected string".to_string())
-            }
-        },
-        namespace: None,
-        to_dsl: None,
-    }
-}
-
-// ========== IPAM types ==========
-
-/// IPAM Pool ID type (e.g., "ipam-pool-0123456789abcdef0")
-pub(crate) fn ipam_pool_id() -> AttributeType {
-    AttributeType::Custom {
-        name: "IpamPoolId".to_string(),
-        base: Box::new(AttributeType::String),
-        validate: |value| {
-            if let Value::String(s) = value {
-                validate_ipam_pool_id(s)
-                    .map_err(|reason| format!("Invalid IPAM Pool ID '{}': {}", s, reason))
-            } else {
-                Err("Expected string".to_string())
-            }
-        },
-        namespace: None,
-        to_dsl: None,
-    }
-}
-
-fn validate_ipam_pool_id(id: &str) -> Result<(), String> {
-    let Some(hex_part) = id.strip_prefix("ipam-pool-") else {
-        return Err("expected format 'ipam-pool-{hex}'".to_string());
-    };
-    if hex_part.len() < 8 {
-        return Err("hex part must be at least 8 characters".to_string());
-    }
-    if !hex_part.chars().all(|c| c.is_ascii_hexdigit()) {
-        return Err("hex part must contain only hex digits".to_string());
-    }
-    Ok(())
-}
-
-// ========== Availability Zone ==========
-
 /// Availability zone type with validation (e.g., "us-east-1a")
-pub(crate) fn availability_zone() -> AttributeType {
+pub fn availability_zone() -> AttributeType {
     AttributeType::Custom {
         name: "AvailabilityZone".to_string(),
         base: Box::new(AttributeType::String),
         validate: |value| {
             if let Value::String(s) = value {
-                // Expect format like "us-east-1a" or DSL format
                 let normalized = extract_enum_value(s).replace('_', "-");
-                // Must end with a single letter (a-z)
-                if let Some(last) = normalized.chars().last()
-                    && last.is_ascii_lowercase()
-                    && normalized.len() > 1
-                    && normalized[..normalized.len() - 1]
-                        .chars()
-                        .all(|c| c.is_ascii_lowercase() || c.is_ascii_digit() || c == '-')
-                {
-                    return Ok(());
-                }
-                Err(format!(
-                    "Invalid availability zone '{}', expected format like 'us-east-1a'",
-                    s
-                ))
+                validate_availability_zone(&normalized)
+                    .map_err(|reason| format!("Invalid availability zone '{}': {}", s, reason))
             } else {
                 Err("Expected string".to_string())
             }
@@ -717,11 +77,9 @@ pub(crate) fn availability_zone() -> AttributeType {
     }
 }
 
-// ========== IAM Policy Document ==========
-
 /// IAM policy document type
 #[allow(dead_code)]
-pub(crate) fn iam_policy_document() -> AttributeType {
+pub fn iam_policy_document() -> AttributeType {
     AttributeType::Map(Box::new(AttributeType::String))
 }
